requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiohttp==3.9.1
pydantic==2.5.2
PyPDF2==3.0.1
PyMuPDF==1.23.8
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0
groq==0.4.1
scikit-learn==1.3.2
httpx==0.27.2
gunicorn==21.2.0
